inheritance draggable report and deleting inherianddepen
