+++ package.json
@@ -28,8 +28,6 @@
   "dependencies": {
     "c3-linearization": "^0.3.0",
     "colors": "^1.3.0",
-    "detect-installed": "^2.0.4",
-    "get-installed-path": "^4.0.8",
     "graphviz": "0.0.8",
     "sha1-file": "^1.0.1",
     "solidity-parser-antlr": "^0.2.15",
