+++ src/inheritance.js
@@ -4,8 +4,10 @@ const fs = require('fs')
 const parser = require('solidity-parser-antlr')
 const graphviz = require('graphviz')
 const { linearize } = require('c3-linearization')
+const moment = require('moment')
+const path = require('path')
 
-export function inheritance(files) {
+export function inheritance(files, options) {
   if (files.length === 0) {
     console.log('No files were specified for analysis in the arguments. Bailing...')
     return
@@ -14,6 +16,9 @@ export function inheritance(files) {
   const digraph = graphviz.digraph('G')
   digraph.set('ratio', 'auto')
   digraph.set('page', '40')
+
+  // for draggable
+  const definition = { "contracts": new Array(), "inheritances": new Array() }
 
   for (let file of files) {
 
@@ -44,6 +49,9 @@ export function inheritance(files) {
           }
 
           digraph.addNode(contractName)
+
+          // for draggable
+          definition.contracts.push(contractName)
         }
 
 
@@ -59,9 +67,15 @@ export function inheritance(files) {
             }
 
             digraph.addNode(dep)
+
+            // for draggalbe
+            definition.contracts.push(dep)
           }
 
           digraph.addEdge(contractName, dep)
+
+          // for draggable
+          definition.inheritances.push(contractName + "=>" + dep)
         }
       }
     })
@@ -69,5 +83,46 @@ export function inheritance(files) {
     dependencies = linearize(dependencies, {reverse: true})
   }
 
-  console.log(digraph.to_dot())
+  if (options.draggable) {
+    try {
+      // generate report
+      const outputFileName = reportGenerate(definition)
+  
+      console.log()
+      console.log("successfully generated!")
+      console.log("open " + outputFileName + " in browser")
+    } catch (e) {
+      console.error(e)
+      process.exit(1)
+    }
+  } else {
+    console.log(digraph.to_dot())
+  }
+}
+
+function reportGenerate(definition) {
+  // remove CR, LE, and space from definition
+  const outputJSON = JSON.stringify(definition).replace(/\r|\n|\s/g, '')
+
+  // load jspulub and jquery
+  const jsplumbDefaultsCss = fs.readFileSync(__dirname + '/../node_modules/jsplumb/css/jsplumbtoolkit-defaults.css').toString()
+  const jsPlumbJs = fs.readFileSync(__dirname + '/../node_modules/jsplumb/dist/js/jsplumb.min.js').toString()
+  const jquery = fs.readFileSync(__dirname + '/../node_modules/jquery/dist/jquery.min.js').toString()
+
+  // load template html
+  const template = fs.readFileSync(__dirname + '/../resources/template.html').toString()
+
+  // generate file name.
+  const m = moment()
+  const timestamp = m.format('YYYYMMDDHHmmss')
+  const outputFileName = 'inheritance_' + timestamp + '.html'
+
+  // generate report
+  let output = template.replace(/{{definition}}/g, outputJSON)
+  output = output.replace(/{{jsplumbtoolkit-defaults.css}}/g, jsplumbDefaultsCss)
+  output = output.replace(/{{jsplumb.min.js}}/g, jsPlumbJs)
+  output = output.replace(/{{jquery.min.js}}/g, jquery)
+  fs.writeFileSync(process.cwd() + path.sep + outputFileName, output)
+
+  return outputFileName
 }
